test: avoid regression on deployment size
The current accepted maximum deployment size is a model of about 2MB
(minus a slight overhead for the other followup records).

We've recently introduced some changes to deployment distribution, and
this should not be affected. Therefore this test is added to the multi
partition deployment tests.

(cherry picked from commit 64c4ddfb0dc945106782f001b1317397037d8540)