fix(transport): avoid out-of-order requests
Avoid out-of-order stream lifecycle requests by using a registration
state machine which handles the state of the remote registration.

The registration keeps track of its remote state (`ADDING`, `ADDED`,
`REMOVING`, `REMOVED`), and local state (`INITIAL` and `CLOSED`), as
well as the single allowed in-flight request for itself. Note that
registration are scoped per host, so we effectively have one
registration per host per aggregated client stream. To know if a stream
is connected to a specific host, we just check if its registration is in
the `ADDED` state.

The `INITIAL` state allows transitioning to `ADDING` or `REMOVED`. An
`ADD` operation can be started when in the `INITIAL` state only.

When that happens, an `ADD` request is sent, and the registration will
transition to `ADDING`. Once a successful `ADD` response is received, it
will transition to `ADDED` if it's still in the `ADDING` state. No
further requests will be sent while it's in the `ADDING` state. When
receiving an `ADD` response, the response is discarded if the state is
not `ADDING` anymore or the request manager is closed.

At any point, a `REMOVE` operation can be started. If it is already in
the `REMOVING` or `REMOVED` states, then nothing happens. If it's in the
`INITIAL` state, then the registration immediately transitions to
`REMOVED` and no request is sent out. If it's in the `ADDING` state,
then it will wait until the request is finished (or times out) before
sending the `REMOVE` request. This is done to reduce the likelihood of
requests being processed out of order on the server side. If it's in the
`ADDED` state, then a request is immediately sent out. Once it receives
a successful response, the registration transitions to `REMOVED`. If
unsuccessful, the request is retried. If at any point it receives a
response but its state is not `REMOVING`, or the request manager is
closed, then the response is discarded.