test: verify that process cache does not get polluted
A bug was reported where a process instance could still be created even
thought the deployment was rejected due to follow-up records exceeding
the record batch size.

This occurred because of a cache being build up with all the processes
that are valid, and this cache is not invalidated once the command is
rejected.

This commit adds a regression test against this scenario.

See: https://github.com/camunda/zeebe/issues/13254