fix: invalidate process cache on deployment rejection
This resolves a critical bug where process instances can be created for
processes from rejected deployments.

The problem was that a cache is build up for each of the processes
during the processing of the Deployment:CREATE command, but this cache
was not invalidated when that command is rejected.

This solution only provides a quick fix by completely clearing the
cache, which comes at some performance loss when deployments are
rejected.

We'll need to investigate other solutions for the long term.

See: https://github.com/camunda/zeebe/issues/13254
(cherry picked from commit 617d6e080b5353ab7b1282e88180b365b2f5507f)

Conflicts:
	engine/src/main/java/io/camunda/zeebe/engine/processing/deployment/DeploymentCreateProcessor.java
	engine/src/main/java/io/camunda/zeebe/engine/state/deployment/ProcessVersionManager.java

Resolution:
        There is no error handling on 8.0, so we have to be a bit more
        explicit about when we want to clear the cache. Additionally,
        the ProcessVersionManager is not a cache but actual state on
        8.0.