fix: invalidate process cache on deployment rejection
This resolves a critical bug where process instances can be created for
processes from rejected deployments.

The problem was that a cache is build up for each of the processes
during the processing of the Deployment:CREATE command, but this cache
was not invalidated when that command is rejected.

This solution only provides a quick fix by completely clearing the
cache, which comes at some performance loss when deployments are
rejected.

We'll need to investigate other solutions for the long term.

See: https://github.com/camunda/zeebe/issues/13254