test(engine): verify StackOverFlow during termination results in banning
This test verifies that when a process instance is terminated and
results in a SO we will ban an instance.
It is not the process instance that got terminated that gets banned!
Instead it's one of the child instance for which the "bubbling up"
caused the SO.

(cherry picked from commit 83f22b06f158fadeda6a78d7f0ed5ec00002c265)