refactor: stabilize container based tests
Stabilize start up time of container based tests by increasing the
concurrency. To keep the minutes down as well, add the new experimental
`--terminate` flag, which will force stop all leased VMs as soon as the
agent is stopped.

We could let the agent keep running until the job exits, but just to be
on the safe side, we'll manually stop it to do a graceful shutdown.