refactor(Dockerfile)!: switch to Alpine based image
Switches the base image of our Dockerfile to an Alpine based container.
To stay up to date with Alpine (as the Temurin builds can be behind by
over a month), we unpack our own JRE into the latest available Alpine
image. We also update all the base packages from Alpine to resolve any
potential security issues.

We evaluated OpenJDK, Azul Zulu, and BellSoft Liberica, as possible JDK
variants. In the end, all are pretty much the same, so we went with Zulu
as they tend to be very up to date, and it's easier to update their
package via their own APK repository.

We build our own custom JRE to further compress the modules, slimming
down the image at the cost of a very small start up increase (~10ms).

> **Note**
> Unfortunately `apk` doesn't let us choose to install only security
updates, but considering the base image contains few packages, we
consider it safe enough to update everything. This will break
reproducibility of builds however, and should be considered a trade-off.

To stay compatible with the Helm chart and SaaS operator, we install
a few missing dependencies in the final image, notably bash. Ideally in
the future this will not be necessary and everyone can switch to `sh`.

BREAKING CHANGE: this may break existing deployment which assume a
Debian or Ubuntu base. Care was taken mostly to be compatible with the
existing Helm charts and SaaS k8s operator.

fix(Dockerfile): ignore Hadolint comment