build: switch to software.xdev.saveactions
> This project has been archived and is looking for a maintainer

This is the disclaimer at the top of the README of
https://github.com/dubreuia/intellij-plugin-save-actions

The project was continued by fishermans in
https://github.com/fishermans/intellij-plugin-save-actions, but they
never published a new version on the IDEA Marketplace.

xdev-software has now done this instead, which is the plugin version
that this commit switches to. Everyone still needs to accept the change
by installing the newly required plugin. On start up, you'll notice an
error pop-up informing you that you have not yet installed a required
plugin.

This project seems nicely set up, with GHA to build and publish the
version to the marketplace. They state that they will provide
maintenance updates to provide support to newer IDEA versions. I had a
look through the diff to look for obvious security problems, but did not
encounter anything suspicious. I think we're good to use this instead.

See: https://github.com/dubreuia/intellij-plugin-save-actions/issues/427
See: https://github.com/fishermans/intellij-plugin-save-actions
See: https://github.com/xdev-software/intellij-plugin-save-actions