feat(engine): write ResourceDeletion DELETING events
Writes the ResourceDeletion DELETING events to the logstream. It also
changes the response to use the DELETING intent. This is to keep DMN
aligned with BPMN, which will be added in the future.

For BPMN the deletion will happen asynchronously. We will respond with
DELETING and won't write DELETED yet.
For DMN we don't have this issue. We can immediately write the DELETED
as a follow-up event.