feat(engine): get process instance keys by process definition
Adds a method to read the PROCESS_INSTANCE_KEY_BY_DEFINITION_KEY
ColumnFamily. It returns a list of process instance keys. These are the
running processes of a specific process definition.

It also returns banned instances. These need to be filtered manually if
a list without banned instances is desired.