fix: group by container and service
It looks like in our current set up we scrape the metrics twice which results in wrong results.
Furthermore, metrics can be reported twice for empty container and actual container.

We exclude the empty container name, and group now by service, container, and pod to see correct results.

Elasticsearch seems to contain two containers in our SaaS environment which is the reason why we want to see the container here as well.