feat(clients/java): allow using custom executors
Enables anyone to pass their own custom executor to the job worker,
which may or may not be owned by the client. If it is owned by the
client, then it is also shut down on client close; otherwise, it's up to
the caller to close it.

The executor will be used to poll for jobs, handle responses, and call
each job handler.

If no custom executor is provided, the old behavior of creating a thread
pool is kept.