fix: mark skipped commands as processed
Previously skipped commands, e.g. because of banned instance have been skipped but caused
to not update the last processed command position. This can lead to issues, if we have
a lot of commands, which need to skip and not update the position we
can't do snapshots, nor delete any data. Furthermore, we need to replay all of them on
restart.

This commit fixes this issue, we mark the command as processed and reset the transient state.
Furthermore we need to commit the state change, so instead directly calling skip we have to run updateState
in order to commit our transaction.