refactor(clients/java): add StreamJobsCommand
Adds a new `StreamJobsCommand` which builds and sends the underlying
`StreamActivatedJobsRequest` using the gateway protocol.

The response is currently an empty response for future proofing and to
keep aligned with other commands. The real response is streamed through
the consumer provided when building the command - i.e. the activated
jobs.

Request time out works a little differently for this command: it is by
default null, and is only added if the user passes one. If provided, it
will cancel the stream upon time out.