fix(clients/java): fixes serialization bug on default value
Protobuf message properties all have default values. When creating an
empty `GatewayOuterClass.ActivatedJob`, for example, it's worker
property will not be null but instead an empty string.

Here, we always assume that `customJobHeaders` is a proper JSON object,
but this is wrong when working with protobuf message - it can also be an
empty string. The fix is just to set an empty map if we detect an empty
string.