feat(engine): add support for scheduled engine tasks
Provides wrappers that should make it easier for the engine to write
scheduled tasks. This bridges a gap between scheduled tasks in the
stream processor and the engine.

By giving scheduled engine tasks a restricted interface, we make it
easier to discover tasks and ensure that they are safe and don't
reuse state.