test(engine): fix flaky test
This test contained a race condition as it was depending on a specific
timing of message correlations. This commit changes this test to use
error events instead. The behavior is the same, but this should make it
deterministic.

(cherry picked from commit b09c141b957bc4b98bb96fe151c163fb996ff4d2)