test(engine): resolve flakiness
This test was using Mockito to verify that the dumpProcessorRef was
called twice before asserting that list of processedInstances.

Unfortunately Mockito is not good with threading. The processRecord
method is called in a separate thread from the assertion. As soon as the
 method is called twice the assertion is true and we continue asserting
 the list. However, because processRecord is in the different thread it
 may not have executed the method yet. This means the list will not
 contain both of the expected elements at time of the assertion.

 We can completely remove the spy. By waiting for 2 COMPLETE_ELEMENT
 commands instead we know that the call to processRecord() is finished
 before we start doing assertions.

(cherry picked from commit ecb743b08123c67d8453d3c4d5522cbb11604e2a)