test(engine): resolve flakiness
We used the getActiveCount + the getCompletedTaskCount. This turns out
to be flaky. According to the javadoc of the getCompletedTaskCount this
returns only an approximation:

"Returns the approximate total number of tasks that have
completed execution. Because the states of tasks and threads
may change dynamically during computation, the returned value
is only an approximation, but one that does not ever decrease
across successive calls."

 By using a mockito spy we can get rid of this approximation, and just
 verify if we called the schedule method instead.