docs: adjust experimental api annotation for zeebe
The annotation was copied from io.grpc.ExperimentalApi and as such still
caried references to the gRPC project. This changes the description to
fit the Zeebe project.

These changes include adjusting the copyright notice at the start of the
license header. As the original copyright is still mentioned in the
file, this should be fine as a copy of the license is still available as
well. This change happens automatically through the
maven-license-plugin.