feat: introduce command with tenant step
This interface can be implemented by any of the command builder
implementations that need to allow the user to explicitly specify a
tenant by its identifier.

The interface is made generically, so it can return the type of builder
that it is added to, so users can chain the calls.

I've spend considerable time to document its usage from a user
perspective, because its usage is highly dependent. This indicates to me
that we're planning to do a lot of magic, which we think is good for the
end user but may simply be unexpected.

I've also tried to highlight that this method is currently experimental
and may not yet have any effect while the multi-tenancy feature is being
developed. Typically, we build features the other way around, but
Connectors and Operate/Tasklist have explicitly requested such an API so
they could be unblocked faster.

I've referenced https://github.com/camunda/zeebe/issues/12653 in the
experimental api as it is the EPIC issue which may provide users a good
overview of the current status of the feature. For change requests, they
can comment there.

I've also marked the api as since 8.3, although that's not something we
typically add. I'm not sure whether we should do this or whether its
better to remove it. I'm open to remove it again.