feat: allow specifying tenant for deploy resources
Adds the experimental method `tenantId` to the deploy resources command
builder. Calling the method has no effect, and the builder is returned
unchanged. This is simply to allow others to start using this method
during development.

Note that the deploy resources command has a slight change in behavior
when the tenantId is not explicitly set by the user. So, I've overridden
the method to extend its javadoc.

At this time, I think this is the only command that will have a
different behavior associated.

The test verifies something odd, but at least it verifies something.
It's mostly there to show that it can be used in practice.