feat: allow specifying tenant for create instance
Adds the experimental method `tenantId` to the create instance command
builder. Calling the method has no effect, and the builder is returned
unchanged. This is simply to allow others to start using this method
during development.

The tests verify something odd, but at least they verify something.
They are mostly there to show that it can be used in practice.

I've added multiple tests because there are multiple ways to reach this
method for the chain. Although the code changes don't reflect this, I
think it makes sense to test the different ways we want users to use
this method.