fix(broker): don't call listeners when transition is cancelled
Instead of completing the transition future successfully, we now
complete exceptionally with a custom exception type.

This ensures that callbacks registered on this future consider the
transition as failed.

With these changes, the `term` and `role` of the partition context are
not updated and partition listeners are not notified about a new role.

(cherry picked from commit b97a2c20adb8447546e6f7cc3ba968447c77cf7d)