refactor(broker): remove transition step context
This removes a mutable context that was updated during transitions to
detect stuck transition steps. This context was available through the
partition context and leaked implementation details of the transition
process.

Instead, we now detect stuck steps by keeping track of the current step
and the time it has started in the `PartitionTransitionProcess`. From
there, it is available to the `ZeebePartitionHealth` without leaking
details to the entire context.

Additionally, the order of reported health issues has changed. On the
first transition, when services have not been installed yet, we report
a stuck transition step instead of the more generic "services not
installed yet" issue.

Co-authored-by: Rodrigo Lopes <132340590+rodrigo-lourenco-lopes@users.noreply.github.com>