fix(broker): always set term and role, even when transition failed
Ensures that cancelled but partially completed transitions are "rolled
back", i.e. services are closed, when preparing the next transition.

For example, when a transition from follower to leader is cancelled but
the stream processor step was already finished, the next time the stream
processor step runs, it must be aware that it has previously
transitioned to leader in order to stop the stream processor again.