fix(broker): only update role and term when prepare phase succeeded
If the prepare phase fails, we assume that none of the steps have
succeeded in preparing for the new role and thus not update role and
term.
For example, if the transition from leader to follower fails in
prepare, the next transition to follower will go through the correct
prepare phase again.
At worst, this will cause a step to go through the same prepare phase
again. This is usually handled correctly on each step (null checks etc.)
and even if not, it will only result in failed transitions and not
in correctness bugs like using the wrong db.

(cherry picked from commit 0d59bb1cfc31818a41c2692e3f4c56686542ff61)