feat: allow specifying tenant for eval decision
Adds the experimental method `tenantId` to the evaluate decision command
builder. Calling this method has no effect, and the builder is returned
unchanged. This is simply to allow others to start using this method
during development.

The tests verify something odd, but at least they verify something. They
are mostly there show that it can be used in practice.

I've added multiple tests because there are multiple ways to reach this
method for the chain. Although the code changes don't reflect this, I
think it makes sense to test the different ways we want users to use
this method.