feat: allow specifying tenant for publish message
Adds the experimental method `tenantId` to the publish message command
builder. Calling this method has no effect, and the builder is returned
unchanged. This is simply to allow others to start using this method
during development.

The test verifies something odd, but at least it verifies something. It
is mostly there to show that it can be used in practice.