feat: introduce command with multiple tenants step
This interface can be implemented by any of the command builder
implementations that need to allow the user to explicitly specify one or
more tenants by their identifiers.

The interface is an extension of CommandWithTenantStep so that it can
inherit the comprehensive documentation of the tenantId(String) method.

Similarly to that interface, I've attempted to highlight the
experimental nature of this interface.

One major difference is that this interface allows chained calls to
`tenantId(String)` to add multiple tenants. This makes sense form the
perspective of this interface as an extension to the original. A similar
style of chaining can be used to build process instance modification
commands (i.e. adding the activate and terminate instructions).

I've also added a shorthand for var args usage, which sometimes is
preferred by users.