fix(engine): delete latest version instead of current version
Latest version is the version number that keeps track of which process
we can start if we start the latest.
Current version is the highest version number that was ever known for a
process.

When we delete the latest version of a process we must make sure the
next time we start a process we start version latest - 1.

Additionally we cannot remove the version information from the state. If
 we do we lose track of the maximum version we have ever deployed for
 this processId.