feat: add owning tenant to deploy resource response
Adds the experimental method `getTenantId` to the response of the deploy
resource command. This only provides a dummy implementation as the
feature is under construction, while allowing others to start using
these methods during their development.

Although marked as experimental, I haven't gone as far with documenting
this as the builder method. My reasoning is that users might wonder how
to use the builer method, while this is simply a getter for data that we
provide.

Please note that the method automatically also becomes available to the
response of the deprecated Deploy Process command. This is because of
re-use between the two commands' interfaces. I don't see this is a
problem. Users should avoid using the deprecated command anyways and
perhaps we finally delete it some day.