feat: add owning tenant to create instance responses
Adds the experimental method `getTenantId` to the response of the create
process instance command and the response of the create process instance
with result command. This only provides a dummy implementation as the
feature is under construction, while allowing others to start using
these methods during their development.

Although marked as experimental, I haven't gone as far with documenting
this as the builder method. My reasoning is that users might wonder how
to use the builder method, while this is simply a getter for data that
we provide.