fix(broker): fail partition if transition preparation fails
This is safer than carrying on and waiting for the next transition to
fix the context and re-install services.

The prepare phase should almost never fail. If it does, it is not clear
which services were re-installed and which not. This undetermined state
is not safe to operate on. For example on snapshot replication, we must
go through the `INACTIVE` role to ensure that the state is re-opened.