refactor(journal): determine segment descriptor length while decoding it
There is no need to keep track of the lengths for each version. Instead sbe can determine
the length while decoding it.

This commit also puts the cur version back to 2, because the addition of new fields is encoded
in sbe by updating the schema version. So now we have version 2 with sbe schema 1 and schema 2.
Schema 1 is the previous version in 8.2.x. Schema 2 is the new version that contains lastPosition
and lastIndex.