feat: add owning tenant to activated job response
Adds the experimental method `getTenantId` to the response of the
activate jobs command, for each job individually. This is relevant
because several jobs can be requested by a single ActivateJobsBatch
command, where they can be requested for multiple tenants at once.

This only provides a dummy implementation as the feature is under
construction, while allowing others to start using these methods during
their development.

Although marked as experimental, I haven't gone as far with documenting
this as the builder method. My reasoning is that users might wonder how
to use the builder method, while this is simply a getter for data that
we provide.