feat(clients/java): opt-in streaming for job worker
Adds opt-in usage of job streams with the job worker. When opted in, on
open, the worker will maintain a long living job stream to a gateway. It
will optionally use a different time out than the default job worker
request time out, as we want polling to be short, but streaming to be
long living.

This has no impact on polling; the idea is that we should be polling for
older jobs as before, and long polling will kick in. Since the brokers
do not perform any form of back filling right now, this is still
necessary for jobs made available before a stream was created.

When the stream is closed - normally or not - it is reopened unless the
worker is closed.