fix(atomix): race condition with multiple snapshot listeners
Fixes a race condition due to multiple, unordered snapshot listeners
which all modify the Raft context asynchronously. The first one was
updating the local snapshot ref, and the second one was compacting the
log. It's important that the snapshot ref is always updated before
compaction occurs, otherwise we may send the wrong prevLogTerm or
prevLogIndex to the follower if we are replicating from the first entry
of the log.

Since compaction is now automatically triggered by the RaftContext, the
RaftRule is updated to make use of this and remove its own custom
compaction logic. This means the tests must be updated to reflect the
conditions when one wants to snapshot awaiting compaction or not.

c

(cherry picked from commit d8f331ac33d36a8b0a533db48e6a016d799daae0)