feat(db): migrate process versions
We need to build a list of known versions for each process version in
our state.
Before the migration we only know what the highest version in the state
it. We know that no version can be deleted yet. So we know we have
version 1 ... highest version. With this information we can create a
loop that loop until it reaches the highest version and add all these
versions to the list.