test(engine): verify process version migration
Adds testcases that verify that process versions are migrated as
expected.
Note: I had to add a getKnownVersions to the NextValue object. This
specifically for this test. It may not be the nicest to add this method
for a test, but I think verifying a succesfull migration warrants this
added method.