test: don't annotate tenant owned as immutable
This fixes failing test
ImmutableProtocolTest.shouldAnnotateImutableProtocol.

All interfaces of record values must be annotated as immutable.

The tenant owned interface is not a record value and does not have to be
annotated as immutable. Like the process instance related interface, we
can exclude it from the immutable protocol test case.