refactor: use pretty json in tests
These tests verify that we can serialize the records to specific json
strings. By using text blocks, we can use prettified json instead.

This does not affect what the tests actually verify, but it makes it
easier to add new properties to the expected json strings when we add
new properties to records. Which I'll be doing in the upcoming commits.