test: recover from call activity exceeds batch
This adds two tests cases where we expect that a call activitity can
recover from exceeding the max batch size limit.

This can happen on activation (creating a child process instance with
variables) and on completion (propagating the variables of the child
process instance).

We want to recover by raising an incident.

Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>