test: resolve incident on call activity exceeding batch
We also want to make sure that we're able to recover from this incident.
To do so, the user will have to override the variables and then resolve
the incident.

Note that we can't add this test case for a completing call activity
where the variables caused the batch size exceeded. The reason for this
is that the child instance is already completed and the variables are
stored internally in an Event Trigger which is not visible in Operate.

We're fine with this for now, since modification could still be used to
recover.

Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>