fix: raise incident on exceeding batch size
This raises an incident on any element instance (except of PROCESS type)
where processing the command results in exceeding the max message size
(i.e. exceeding the batch size limitation).

The incident should be resolvable in pretty much all cases. There is a
test for call activity activation explicitly.

When the incident cannot be resolved (i.e. it is resolved and a new
incident is raised), then you can use modification to recover.

This is in all cases better than the previous situation where such
process instances would be banned.

Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>