fix: only disable job on job incidents
When an incident is created we disable the job in the state directly
from the incident created event applier (which is ugly, but okay).

Since we now have a new case where the MESSAGE_SIZE_EXCEEDED error type
is used we should deal with this. The job should simply only be disabled
when a job is referred to from the incident. Otherwise, if the incident
did not refer to a job, then there is no job to disable anyways.

Co-authored-by: Remco Westerhoud <remco@westerhoud.nl>
Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>
(cherry picked from commit 3bd36781fab025c3ba04983bc17ecf6215c75a48)