feat: use recognizable terminology in incident
Users are more familiar with processing an element than transitioning.

Likewise, ran into is also not really understandable. What happened is
that the MAX_MESSAGE_SIZE limit was exceeded. When users search for this
term in the forums, they find many helpful posts. And our docs simply
show the configuration options.