test(qa): verify process version after upgrade
Adds a test scenario to verify the deployed version is incremented as
expected after an upgrade.
These test cases are really supposed to work with process instances.
Therefore I had to make sure to deploy a process and start an instance
for it to not fail on checking if the process instance completed. It
still verifies the version of a redeployment correctly.