refactor(engine): move method call
Moves the marking the migration as finished to the handleMigrationTask
method. This way it's more obvious that this happens when it ran, or
when it skipped.

The method is now also inlined so it's clear it is marked in the
migration state.

(cherry picked from commit b310397e1cc2653fc31e9ca8e9e4a8925bf02349)