refactor(engine): needsToRun can be true by default
New migration should always use the new ColumnFamily to determine if
they need to run. As we already check if the migration ran before we
execute this needsToRun check we can return true by default.

For older migrations we cannot make this assumption. They have already
ran, but are not marked as FINISHED in the state yet. Therefore they
still override this method and use their old implementation. Once
they've been skipped once, they are part of the state and will be
skipped by checking if it's FINISHED in the future.

(cherry picked from commit d2694720e2a035384b6905cca11e4d200c624677)