refactor(dist): reduce number of Spring threads
Greatly reduce the number of Spring threads by:

- Disabling auto configuration for the reactive `WebClient` (we don't
  use it). This means we cannot autowire `WebClient` beans, but as I
  said, we don't use it.
- Reducing the default cap for the bounded elastic scheduler (used by
  the actuators to schedule tasks, e.g. delayed health indicators) from
  `10 * numberOfCores` to `2 * numberOfCores`

Additionally remove the quiet period to shutdown the management server.
This may be useful when it is a proper server handling user requests,
but I think it's acceptable for a management server to simply shut off
immediately. This will prevent unnecessary pauses of up to 2s when
shutting down the application.