feat(clients/java): introduce job worker metrics
Adds a new API to the job worker builder that allows users to pass a
custom `JobWorkerMetrics` implementation. This currently supports two
metrics:

- `jobActivated(int count)`: called every time one or more jobs is
  activated (but not yet handled)
- `jobHandled(int count)`: called every time one or more jobs has been
  handled by the `JobHandler`

> **Note**
> While we currently always call the above for each individual jobs, I
opted to pass in the count for future proofing.

Additionally, an optional implementation was added for Micrometer, which
will help integration with Spring and Micronaut. Users can call
`JobWorkerMetrics#micrometer`, which returns a builder that lets them
pass a custom `MeterRegistry` and a set of fixed tags (e.g. the worker
job type, the app version, etc.)