feat(zeebe-client-java): new API to allow user to get single variable by name
Added a new getVariable(name) API in ActivatedJob and ProcessInstanceResult interfaces to allow user to get only a single variable by name/key. This reduce the boilerplate of doing job.getVariablesAsMap().get(name).
It is introduced also an in memory cache for the variables, instead of every time deserialize the variables, now the deserialization is made only the first time.