refactor: extract decisionmetadata interface
We need to make the DecisionRecord TenantOwned. However, if we do that,
then we also adjust the deployment record indirectly because it contains
decisionMetadata. That decisionMetadata field is directly using the
DecisionRecord, so then we add tenantId into each entry of the
decisionMetadata field. This is unnecessary and unexpected because the
tenant id is already set for the entire deployment. Individual resources
of a single deployment cannot be owned by different tenants.

So, we can extract the DecisionMetadata interface. This allows us to use
this in the DeploymentRecord, while adding tenant id to the
DecisionRecord.