refactor: use decision metadata in deployment record
This is a continuation of the previous commit. By using decision
metadata in the deployment record, we can make the decision record
tenant owned without adding the tenant id to the decision metadata in
the deployment.

Note that this refactor requires some small changes in tests where the
explicit DecisionRecordValue interface was used when testing
deployments. Instead, we now have to use DecisionMetadataValue. Such a
change is breaking backward compatibility of the protocol api, but only
in those places where the specific returned value of the deployment
record's getDecisionMetadata is stored in an explicitly typed variable.

Since the protocol is not part of the public api, we are allowed to make
such changes. However, we should document it in the update guide.