revert: extract decision metadata value
We should not break backward compatibility of the protocol if possible.

We're chosing here between breaking backward compatibility of the type
of one of the Deployment record properties, and not having tenant id
duplicated in the metadata of the deployed resources.

The benefit of removing this duplication does not outway breaking
backward compatibility.

Therefore:

This reverts commit a04414d5d66f005b8f5152d6ac8b96e20eeff4f7.
This reverts commit d1143e5aeca680016a0836d4fbc6f35dcbdb6d61.
This reverts commit 7d206b45526bb29f63b23f4b05c7f3d0a7386950.
This reverts commit 539d88491b33ea1ae61b99dbf61a17a481358209.