feat: allow job batch records to specify tenant ids
For job batch records we need the ability to set one or more tenant ids.

Similar to the CommandWithOneOrMoreTenantsStep this can be a list of
strings.

I've chosen not to extract this into a separate interface, as I don't
expect that we'll have many other record values that require something
similar. If we do, we can add this as a unique method on each.