feat(engine): cache DRGs in memory
During evaluation of a Decision we need to have the parsed DRG
available. We used to parse is everytime it was required. This had a
significant impact on the performance of business rule tasks.

With this commit we parse the DRG only once upon retrieving it. The
parsed DRG is then cached in memory. Everytime it is retrieved it's
taken from the cache if it exists in there. Otherwise, the CacheLoader
retrieves it from the ColumnFamily.

The cache has a maximum size. This is set to 10k for now (will be
configurable in a later commit). This means that a maximum of 10k DRGs
can be stored in the cache. When the limit is exceeded the cache will
evict the DRGs that are used the least. This should result in the latest
 version of DRGs always being available in the cache.

(cherry picked from commit e3edfc54f0609fb40be90075703754df94378ba4)