feat(engine): resubscribe timer start events
When the latest version of a process is deleted we must reactivate the
timers of the previous version. This means we need to write the Timer
CREATED event for all timer start events in the process.

With this commit we find the previous version, find the process that
belongs to it and write the CREATED event for any timer start event in
the process.