fix: reduce timer resolution to ms
Previously we converted the given delay to nanos which caused issues when the delay was over 290+ years.
Of course the such delay in general doesn't make much sense, but also the converting to nanos was questionable.
We convert later, when scheduling the timer, the time to milliseconds since the deadtimewheel we use
or implement only supports milliseconds.

This commit changes the delay to nano conversation to milliseconds.