feat(broker): introduce new gossiper for disseminating cluster topology
We are not re-using SWIM mainly for separation of concerns. SWIM gossip
is not designed to gossip a shared mutable state, rather only to send
properties of a each member. ClusterTopology is a shared state which
any member can update. So we would like to model it as a shared state.
Although we can use SWIM to disseminate this information, it is not the
optimal way of handling it. Moreover, adding a new implementation lets us
move to protobuf with out breaking compatibility.