feat(topology): add protobuf messages for gossipping topology updates
We decided to use protobuf to serialize ClusterTopology. We don't want
to use schema-less formats like Json. Using SBE is not really necessary
here as zero-copy is no that useful. We eventually want to move to gRPC,
so starting with protobuf might already help.

The protobuf message are prefixed with M to distinguish from the corresponding
internal records.