ci(build): add maven wrapper
This adds maven-wrapper to our repository, allowing for a central place
to configure which maven version we are using.
This is helpful for local development as well as CI where we usually
have to be careful to use the "right" maven version and remember to
upgrade regularly.

Late backport of 85c83e9f05c271906a8247dbf94e7692915b57b6