feat(zb-db): allow opening DBs for snapshots only
Adds a new API call to `ZeebeDbFactory` which allows opening existing
databases in a snapshot-only mode. This returns a `ZeebeDb` which can
ONLY take snapshots, and will throw a `new
UnsupportedOperationException` for any other operations.

While it would be nice to have a proper read-only DB implementation,
there are no transactions in a such a DB, and our whole access to the DB
is coupled with the transaction API. As we were thinking of switching
away from transactions, it will likely be easier after that to implement
a read-only DB.