refactor(broker): speed up recovery of state controller
Speeds up the recover of the state controller by creating a snapshot out
of the latest snapshot for the runtime. This greatly improves
performance when the runtime and snapshot directories are on the same
filesystem, as RocksDB will create hard-links for all the SST files
instead of actually copying the data.