feat(auth): add a tenant access checker class
Add a TenantAccessChecker interface and it's implementation.
The purpose of this class is to provide helper methods for
checking if a provided list of authorized tenants contains
a provided tenant id (or list of ids).

This class will be used to validate if a request is authorized
to perform changes on the data of a given tenant.