feat(topology): coordinator queries known members before generating topology
If a coordinator restarts with a data loss, then its local persisted topology will be
empty. If the coordinator generates the topology from static configuration, there is
a possibility of inconsistent topology if the static configuration does not match
the actual cluster topology.

To prevent this, coordinator first queries known members before generating the topology.
If at least one member return an uninitalized topology, coordinator assumes that this is
first bootstrap of the cluster and generates a new topology.