fix(engine): propagate only defined variables
This commit introduces a new property called
`propagateAllParentVariables` to allow a user to control if all
variables or just specific ones need to be propagated to a child process
instances when activating a call activity.

1. (Default) `propagateAllParentVariables = true`, then Zeebe behaves as
currently (i.e., all variables a propagated to the child process
instance).
2. `propagateAllParentVariables = false`, only the variables defined via
input mappings are propagated to the child process instance (if any).

The property `propagateAllParentVariables` can be set per call activity.