feat(protocol): add an authorization field to record metadata
Add an authorization field to the RecordMetadata SBE message of the Zeebe protocol.
Initially,the authorization field will be populated only for Command records, and
this data won't be exported. The field will contain information on the authorizations
of the user that made the client request that created the Command record. The field type
is of variable length, as authorizations may vary per user.