feat(protocol-impl): add metadata authorization implementation
Add the AuthInfo class, which is an implementation for the authorization field of
Record Metadata. The class contains format and authData properties. The authData
property is encoded into a String value, and the format property provides information
on what mechanism encoded/decodes the authData value. Initially, only the JWT format
is available, but new ones can easily be added (ex. msg-pack).