fix(broker): remove existing limiter before adding a new one
The previous assumption was that onBecomingFollower is always invoked before becoming leader again.
But recent changes to ZeebePartition resulted in not notifying the listeners, if a transition was
cancelled in between. When a Leader -> Follower -> Leader transition happened, where follower transitioned
was cancelled, then the command api uses the limiter from the previous leader role. To prevent this, when
adding a partition the limiter force create a new one. addPartiton is only used when we want to re-add
a partition. So it makes sense to force re-create it.

(cherry picked from commit f26961fe8c3c50ec2998953f8ca79f942f3107b4)