fix(raft): use fixed algorithm to calculate member hash
This replaces a call to a known buggy version of murmur32 with the
fixed version. The fixed version will return different results only if
the member id contained unusual characters such as emoji.
Since our member ids are usually just numbers, and the hash appears to
be unused anyway, we can accept this change.