refactor(dist): register only non-auto configured metrics
Simplifies the metrics auto-configuration and registration for all
applications.

First, removes registration of all metrics which can be auto-configured
by Spring and Micrometer. This paves the way to eventually use
Micrometer in the application as a metrics facade instead of Prometheus
directly.

Additionally, register Netty allocator metrics which are not counted in
the total direct memory bounds of the JVM.